RI_Stack/java/src/base/org/dvb/media/BackgroundVideoPresentationControl.java
// COPYRIGHT_BEGIN
//  DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER
//  
//  Copyright (C) 2008-2013, Cable Television Laboratories, Inc. 
//  
//  This software is available under multiple licenses: 
//  
//  (1) BSD 2-clause 
//   Redistribution and use in source and binary forms, with or without modification, are
//   permitted provided that the following conditions are met:
//        ·Redistributions of source code must retain the above copyright notice, this list 
//             of conditions and the following disclaimer.
//        ·Redistributions in binary form must reproduce the above copyright notice, this list of conditions 
//             and the following disclaimer in the documentation and/or other materials provided with the 
//             distribution.
//   THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS 
//   "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED 
//   TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A 
//   PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT 
//   HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, 
//   SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT 
//   LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, 
//   DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY 
//   THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT 
//   (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF 
//   THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
//  
//  (2) GPL Version 2
//   This program is free software; you can redistribute it and/or modify
//   it under the terms of the GNU General Public License as published by
//   the Free Software Foundation, version 2. This program is distributed
//   in the hope that it will be useful, but WITHOUT ANY WARRANTY; without
//   even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR
//   PURPOSE. See the GNU General Public License for more details.
//  
//   You should have received a copy of the GNU General Public License along
//   with this program.If not, see<http:www.gnu.org/licenses/>.
//  
//  (3)CableLabs License
//   If you or the company you represent has a separate agreement with CableLabs
//   concerning the use of this code, your rights and obligations with respect
//   to this code shall be as set forth therein. No license is granted hereunder
//   for any other purpose.
//  
//   Please contact CableLabs if you need additional information or 
//   have any questions.
//  
//       CableLabs
//       858 Coal Creek Cir
//       Louisville, CO 80027-9750
//       303 661-9100
// COPYRIGHT_END

package org.dvb.media;

/**
 * A control to support the setting and querying of the video presentation for
 * background players.
 */

public interface BackgroundVideoPresentationControl extends VideoPresentationControl
{
    /**
     * Sets a new video transformation (clipping/scaling/positioning). If the
     * new video transformation is not supported, then the video transformation
     * will not be changed at all (no best effort attempt is made).
     * 
     * @param t
     *            the new video transformation
     * 
     * @return true if the video transformation is supported and has been set,
     *         false otherwise.
     */
    public abstract boolean setVideoTransformation(VideoTransformation t);

    /**
     * Return the current video transformation
     * 
     * @return the video transformation (clipping/scaling/positioning) that is
     *         currently used for displaying the video.
     */
    public abstract VideoTransformation getVideoTransformation();

    /**
     * This method takes a video transformation and returns the closest match of
     * that video transformation that can be supported for the currently
     * selected video. If the input video transformation can be supported, then
     * the output video transformation will have the same parameters as the
     * input video transformation. The definition of 'closest match' is
     * implementation dependent.
     * 
     * @param t
     *            the input video transformation
     * 
     * @return the closest match to the input video transformation. If the input
     *         video transformation is supported, then the input video
     *         transformation will be returned (the same instance), otherwise a
     *         newly created instance will be returned.
     */
    public abstract VideoTransformation getClosestMatch(VideoTransformation t);
}
